hw_3/requirements.txt
pytest
pytest-xdist
selenium
allure-pytest
webdriver-manager
Faker
requests